+++ commands/fun/reverse.js
@@ -23,9 +23,6 @@ module.exports = {
       return message.channel.send(`**${message.author.username}**, you need to give me a message to reverse!`);
     }
 
-    // get everything after the message, join it as one combined string, then reverse it
-    let text = args.join(" ").split("").reverse().join("");
-
     // check if bot has "manage messages" permissions
     if (message.guild.me.permissionsIn(message.channel).has("MANAGE_MESSAGES")) { message.delete(); }
 
@@ -47,6 +44,8 @@ module.exports = {
       }
     }
 
+    // get everything after the message, join it as one combined string, then reverse it
+    let text = args.join(" ").split("").reverse().join("");
     return message.channel.send(text);
   }
 }